# opt: try predict container size using SIMD