--- a/internal/decoder/stubs_go116.go
+++ b/internal/decoder/stubs_go116.go
@@ -80,6 +80,9 @@ func growslice(et *rt.GoType, old rt.GoSlice, cap int) rt.GoSlice
 //go:linkname makemap_small runtime.makemap_small
 func makemap_small() unsafe.Pointer
 
+//go:linkname makemap runtime.makemap
+func makemap(t *rt.GoType, cap int, h unsafe.Pointer) unsafe.Pointer
+
 //go:linkname mapassign runtime.mapassign
 //goland:noinspection GoUnusedParameter
 func mapassign(t *rt.GoType, h unsafe.Pointer, k unsafe.Pointer) unsafe.Pointer