… .github/workflows/unit_test-linux-x64.yml
… internal/decoder/assembler_test.go
… internal/decoder/decoder_test.go
… internal/decoder/generic_test.go
… internal/native/avx/native_amd64.go
… internal/native/avx/native_amd64_test.go
… internal/native/avx/native_subr_amd64.go
… internal/native/avx/native_text_amd64.go
… internal/native/avx/recover_amd64_test.go
… internal/native/avx2/native_amd64.go
… internal/native/avx2/native_amd64_test.go
… internal/native/avx2/native_subr_amd64.go
… internal/native/avx2/native_text_amd64.go
… internal/native/avx2/recover_amd64_test.go
… internal/native/dispatch_amd64.go
… internal/native/native_amd64.tmpl
… internal/native/native_amd64_test.tmpl
… internal/native/recover_amd64_test.tmpl
… internal/native/sse/native_amd64.go
… internal/native/sse/native_amd64_test.go
… internal/native/sse/native_subr_amd64.go
… internal/native/sse/native_text_amd64.go
… internal/native/sse/recover_amd64_test.go
… native/native.h
… native/scanning.c
… option/init.go
… option/option.go